ScooterRental CalculateIncome logic transferred to RentalCalculator
